8d03048d45696, update tpu-bmodel and tpu-runtime
1. support io-reloc mode
2. add runtime flag BM_RUNTIME_SOC_CMDBUF_MEM

+++ driver/bm_bgm.h
@@ -45,7 +45,7 @@ typedef struct bm_mem_desc {
 	} u;
 
 	bm_mem_flags_t         flags;
-	unsigned long                    size;
+	unsigned int                    size;
 } bm_mem_desc_t;
 
 typedef struct bm_mem_desc_u64 {
